Add files via upload
The code that is offered includes a mechanism for file synchronization over MQTT and attendance tracking based on facial recognition. It tracks attendance, takes pictures of faces, and compares them to a database. It also uses MQTT to synchronize a folder's contents between a Raspberry Pi and a laptop. Effective attendance control and real-time file sharing across devices are made possible by this solution.